# api/openapi.yaml
openapi: 3.1.0
info:
  title: Mirrormere API
  description: Foundation API specification for the Mirrormere ambient smart display daemon.
  version: 0.1.0
paths:
...
  /api/widgets/{widget_id}/render:
    get:
      summary: Render widget HTML fragment
      description: Retrieves the server-rendered HTML fragment for a single widget instance on demand.
      operationId: getWidgetRender
      parameters:
        - name: widget_id
          in: path
          required: true
          description: Unique identifier of the widget instance
          schema:
            type: string
      responses:
        '200':
          description: Server-rendered HTML fragment
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Widget not found in active configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/widgets/{widget_id}/push:
    post:
      summary: Push realtime widget payload
      description: Ingests an immediate domain data payload for a custom HTTP provider widget, validates against manifest response_schema, updates SWR cache, and broadcasts widget.update over SSE.
      operationId: postWidgetPush
      parameters:
        - name: widget_id
          in: path
          required: true
          description: Unique identifier of the target widget instance
          schema:
            type: string
      requestBody:
        required: true
        description: Custom widget domain data object matching manifest response_schema
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Push payload accepted and broadcast emitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WidgetPushResponse'
        '400':
          description: Invalid JSON syntax or failed response_schema validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Widget not found in active configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Target widget uses a non-HTTP provider (e.g. calendar, weather, photos, spacer, or tasks) which cannot accept push updates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /widget-types/{type}/assets/{path}:
    get:
      summary: Serve static widget package asset
      description: Serves static assets from the resolved widget package directory with path traversal protection.
      operationId: getWidgetAsset
      parameters:
        - name: type
          in: path
          required: true
          description: Widget type identifier
          schema:
            type: string
        - name: path
          in: path
          required: true
          description: Relative path to the asset within the widget package assets directory
          schema:
            type: string
      responses:
        '200':
          description: Static asset file stream
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '400':
          description: Invalid path or path traversal attempt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/screen/select:
    post:
      summary: Select screen by index
      description: Selects an active screen directly by zero-based index and broadcasts screen.rotate.
      operationId: postScreenSelect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScreenSelectRequest'
      responses:
        '200':
          description: Screen selected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreenNavigationResponse'
        '400':
          description: Invalid screen index
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/screen/advance:
    post:
      summary: Advance screen sequentially
      description: Advances the active screen forward or backward and broadcasts screen.rotate.
      operationId: postScreenAdvance
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScreenAdvanceRequest'
      responses:
        '200':
          description: Screen advanced successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreenNavigationResponse'
        '400':
          description: Invalid direction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/screen/pause:
    post:
      summary: Pause or resume automatic rotation
      description: Controls the automatic screen rotation timer loop with optional auto-resume duration.
      operationId: postScreenPause
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScreenPauseRequest'
      responses:
        '200':
          description: Screen rotation pause state updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScreenPauseResponse'
        '400':
          description: Invalid pause request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/lists/{list_id}/items:
    get:
      summary: Get household list items
      description: Retrieves all current items for a specific list from the local store with optional completed filter.
      operationId: getListItems
      parameters:
        - name: list_id
          in: path
          required: true
          description: Canonical list identifier
          schema:
            type: string
        - name: include_done
          in: query
          required: false
          description: If false, filters out completed items (default true)
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: List items retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListItem'
        '400':
          description: Invalid query parameter syntax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: List not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/audio/volume:
    post:
      summary: Set master audio volume
      description: Sets the master audio volume percentage (0-100) and broadcasts audio.state.
      operationId: postAudioVolume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AudioVolumeRequest'
      responses:
        '200':
          description: Audio volume updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudioStateResponse'
        '400':
          description: Invalid volume request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/audio/mute:
    post:
      summary: Set or toggle master audio mute
      description: Sets or toggles the master audio mute state and broadcasts audio.state.
      operationId: postAudioMute
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AudioMuteRequest'
      responses:
        '200':
          description: Audio mute state updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AudioStateResponse'
        '400':
          description: Invalid mute request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/video/trigger:
    post:
      summary: Trigger video stream
      description: Mounts or updates a video stream in the priority stack and broadcasts video.state.
      operationId: postVideoTrigger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoTriggerRequest'
      responses:
        '200':
          description: Video stream triggered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoStateResponse'
        '400':
          description: Invalid video trigger request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/video/dismiss:
    post:
      summary: Dismiss video stream
      description: Dismisses an active video stream from primary or PiP and broadcasts video.state.
      operationId: postVideoDismiss
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoDismissRequest'
      responses:
        '200':
          description: Video stream dismissed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoStateResponse'
        '400':
          description: Invalid video dismiss request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stream not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/video/action:
    post:
      summary: Dispatch video transport action
      description: Forwards remote playback action to active stream control_url.
      operationId: postVideoAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoActionRequest'
      responses:
        '200':
          description: Video action dispatched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
        '400':
          description: Invalid video action request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stream not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Stream is not controllable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Stream controller unreachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/video/state:
    get:
      summary: Get video presentation state
      description: Returns the active video presentation mode and stream assignments.
      operationId: getVideoState
      responses:
        '200':
          description: Video state retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoStateResponse'
    post:
      summary: Update video player transport state
      description: Webhook dispatched by sidecar when player state mutates and broadcasts video.state.
      operationId: postVideoState
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoPlayerStateRequest'
      responses:
        '200':
          description: Video player state updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoPlayerStateResponse'
        '400':
          description: Invalid player state request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Stream not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/voice/state:
    post:
      summary: Update voice interaction lifecycle state
      description: Relays voice interaction state from voice companion to core and broadcasts voice.state.
      operationId: postVoiceState
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoiceStateRequest'
      responses:
        '200':
          description: Voice state updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoiceStateResponse'
        '400':
          description: Invalid voice state request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ScreenSelectRequest:
      type: object
      required:
        - screen_index
      properties:
        screen_index:
          type: integer
          description: 0-indexed screen number
          example: 1
    ScreenAdvanceRequest:
      type: object
      properties:
        direction:
          type: string
          enum: [next, prev]
          default: next
          description: Direction to advance screen (next or prev)
          example: next
    ScreenNavigationResponse:
      type: object
      required:
        - status
        - current_screen
        - total_screens
      properties:
        status:
          type: string
          example: ok
        current_screen:
          type: integer
          example: 1
        total_screens:
          type: integer
          example: 2
    ScreenPauseRequest:
      type: object
      required:
        - paused
      properties:
        paused:
          type: boolean
          description: true to pause automatic rotation; false to resume
          example: true
        duration_seconds:
          type: integer
          minimum: 0
          description: Seconds to suspend rotation before auto-resuming (0 = indefinitely)
          example: 120
    ScreenPauseResponse:
      type: object
      required:
        - status
        - paused
        - current_screen
      properties:
        status:
          type: string
          example: ok
        paused:
          type: boolean
          example: true
        current_screen:
          type: integer
          example: 0
...
    ErrorResponse:
      type: object
      required:
        - status
        - error
      properties:
        status:
          type: string
          example: error
        error:
          type: string
          example: widget 'custom-sensor-hud' not found in active configuration
    WidgetPushResponse:
      type: object
      required:
        - status
        - widget_id
        - updated_at
      properties:
        status:
          type: string
          example: ok
        widget_id:
          type: string
          example: custom-sensor-hud
        updated_at:
          type: string
          format: date-time
          example: 2026-09-24T22:20:00Z
    ListItem:
      type: object
      required:
        - id
        - list_id
        - title
        - done
        - position
        - created_at
        - updated_at
      properties:
        id:
          type: string
          example: i1
        list_id:
          type: string
          example: groceries
        title:
          type: string
          example: Oat milk
        done:
          type: boolean
          example: false
        section:
          type: string
          example: Dairy
        position:
          type: integer
          example: 0
        assignee:
          type: string
          nullable: true
          example: Alex
        due_date:
          type: string
          nullable: true
          example: "2026-09-26"
        created_at:
          type: string
          format: date-time
          example: 2026-09-24T22:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2026-09-24T22:30:00Z
    AudioStateResponse:
      type: object
      required:
        - status
        - volume
        - muted
      properties:
        status:
          type: string
          example: ok
        volume:
          type: integer
          minimum: 0
          maximum: 100
          example: 75
        muted:
          type: boolean
          example: false
    AudioVolumeRequest:
      type: object
      required:
        - volume
      properties:
        volume:
          type: integer
          minimum: 0
          maximum: 100
          example: 75
    AudioMuteRequest:
      type: object
      properties:
        muted:
          type: boolean
          example: true
...